refactor(build): extract ./constants to /libs/env
